remove SceneNode and Transform
