+++ BasicRenderer/include/MathUtils.h
@@ -1,9 +1,10 @@
 #pragma once
+#include <DirectXMath.h>
 
 #include "Components.h"
-#include "Transform.h"
+#include "MovementState.h"
 
 void ApplyMovement(Components::Position& pos, const Components::Rotation& rot, const MovementState& movement, float deltaTime);
 void RotatePitchYaw(Components::Rotation& rot, float pitch, float yaw);
-XMVECTOR GetForwardFromMatrix(const DirectX::XMMATRIX& matrix);
-XMVECTOR GetUpFromMatrix(const DirectX::XMMATRIX& matrix);
+DirectX::XMVECTOR GetForwardFromMatrix(const DirectX::XMMATRIX& matrix);
+DirectX::XMVECTOR GetUpFromMatrix(const DirectX::XMMATRIX& matrix);
