+++ BasicRenderer/include/MovementState.h
@@ -1,0 +1,10 @@
+#pragma once
+
+struct MovementState {
+    float forwardMagnitude = 0.0;
+    float backwardMagnitude = 0.0;
+    float rightMagnitude = 0.0;
+    float leftMagnitude = 0.0;
+    float upMagnitude = 0.0;
+    float downMagnitude = 0.0;
+};
