+++ BasicRenderer/src/AnimationController.cpp
@@ -1,6 +1,4 @@
 #include "AnimationController.h"
-#include "Transform.h"
-#include "SceneNode.h"
 
 AnimationController::AnimationController()
     : animationClip(nullptr), currentTime(0.0f), isPlaying(true) {
