+++ BasicRenderer/src/DefaultDirection.cpp
@@ -1,0 +1,3 @@
+#include "DefaultDirection.h"
+
+DirectX::XMVECTOR defaultDirection = DirectX::XMVectorSet(0.0f, 0.0f, 1.0f, 0.0f);
